move code to server/query/mpp
