--- cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/common/FillPolicy.java
@@ -1,5 +1,0 @@
-package org.apache.iotdb.cluster.query.distribution.common;
-
-public enum FillPolicy {
-    PREVIOUS,
-}
