--- cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/common/WithoutPolicy.java
@@ -1,6 +1,0 @@
-package org.apache.iotdb.cluster.query.distribution.common;
-
-public enum WithoutPolicy {
-    CONTAINS_NULL,
-    ALL_NULL
-}
