--- cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/plan/process/ProcessNode.java
@@ -1,11 +1,0 @@
-package org.apache.iotdb.cluster.query.distribution.plan.process;
-
-import org.apache.iotdb.cluster.query.distribution.common.TsBlock;
-import org.apache.iotdb.cluster.query.distribution.plan.PlanNode;
-import org.apache.iotdb.cluster.query.distribution.plan.PlanNodeId;
-
-public class ProcessNode extends PlanNode<TsBlock> {
-    public ProcessNode(PlanNodeId id) {
-        super(id);
-    }
-}
