+++ server/src/main/java/org/apache/iotdb/db/query/mpp/common/FillPolicy.java
@@ -1,0 +1,5 @@
+package org.apache.iotdb.db.query.mpp.common;
+
+public enum FillPolicy {
+    PREVIOUS,
+}
