+++ server/src/main/java/org/apache/iotdb/db/query/mpp/common/WithoutPolicy.java
@@ -1,0 +1,6 @@
+package org.apache.iotdb.db.query.mpp.common;
+
+public enum WithoutPolicy {
+    CONTAINS_NULL,
+    ALL_NULL
+}
